+++ app/PageBuilder/Gutenberg/Blocks/Detail.php
@@ -102,67 +102,11 @@ class Detail extends Blocks_Basis {
 			return '';
 		}
 
-		// set actual language.
-		$position->set_lang( Languages::get_instance()->get_current_lang() );
-
-		// get setting for colon.
-		$colon = ': ';
-		if ( false === $attributes['colon'] ) {
-			$colon = '';
-		}
-
-		// get setting for line break.
-		$line_break = '<br>';
-		if ( false === $attributes['wrap'] ) {
-			$line_break = '';
-		}
-
-		// get separator.
-		$separator = get_option( 'personioIntegrationTemplateExcerptSeparator' ) . ' ';
-		if ( ! empty( $attributes['separator'] ) ) {
-			$separator = $attributes['separator'];
-		}
-
-		// get settings for templates.
-		$template = 'default';
-		if ( ! empty( $attributes['template'] ) ) {
-			$template = $attributes['template'];
-		}
-
-		/**
-		 * Filter the attributes for the output.
-		 *
-		 * @since 3.0.0 Available since 3.0.0
-		 * @param array $attribute_defaults The parameter we use.
-		 * @param array $attributes The attributes from PageBuilder.
-		 */
-		$attributes = apply_filters( 'personio_integration_get_list_attributes', $attributes, $attributes );
-
-		// collect the details in this array.
-		$details = array();
-		$taxonomy_data = array();
-
-		// loop through the chosen details.
-		foreach ( $attributes['excerptTemplates'] as $detail ) {
-			// get the terms of this taxonomy.
-			foreach ( Taxonomies::get_instance()->get_taxonomies() as $taxonomy_name => $taxonomy ) {
-				if ( $detail === $taxonomy['slug'] ) {
-					// get value.
-					$value = $position->get_term_by_field( $taxonomy_name, 'name' );
-
-					// bail if no value is available.
-					if ( empty( $value ) ) {
-						continue;
-					}
-
-					// get labels of this taxonomy.
-					$labels = Taxonomies::get_instance()->get_taxonomy_label( $taxonomy_name );
-
-					$details[ $labels['name'] ] = $value;
-					$taxonomy_data[ $labels['name'] ] = get_taxonomy( $taxonomy_name );
-				}
-			}
-		}
+		// map the settings.
+		$attributes['excerpt'] = $attributes['excerptTemplates'];
+		$attributes['lang'] = Languages::get_instance()->get_current_lang();
+		$attributes['line_break'] = $attributes['wrap'];
+		$attributes['excerpt_template'] = $attributes['template'];
 
 		// get block-classes.
 		if ( function_exists( 'get_block_wrapper_attributes' ) ) {
@@ -171,7 +115,7 @@ class Detail extends Blocks_Basis {
 
 		// get content for output.
 		ob_start();
-		include Templates::get_instance()->get_template( 'parts/details/' . $template . '.php' );
+		Templates::get_instance()->get_excerpt( $position, $attributes );
 		$content = ob_get_clean();
 		if( ! $content ) {
 			return '';