+++ app/PersonioIntegration/PostTypes/PersonioPosition.php
@@ -23,12 +23,14 @@ use PersonioIntegrationLight\PersonioIntegration\Post_Type;
 use PersonioIntegrationLight\PersonioIntegration\Statistics;
 use PersonioIntegrationLight\PersonioIntegration\Taxonomies;
 use PersonioIntegrationLight\PersonioIntegration\Themes;
+use PersonioIntegrationLight\PersonioIntegration\Widgets;
 use PersonioIntegrationLight\Plugin\Admin\Admin;
 use PersonioIntegrationLight\Plugin\Compatibilities\Loco;
 use PersonioIntegrationLight\Plugin\Compatibilities\SayWhat;
 use PersonioIntegrationLight\Plugin\Languages;
 use PersonioIntegrationLight\Plugin\Setup;
 use PersonioIntegrationLight\Plugin\Templates;
+use \PersonioIntegrationLight\PersonioIntegration\Widgets\Archive;
 use WP_Post;
 use WP_Query;
 use WP_REST_Request;
@@ -92,6 +94,9 @@ class PersonioPosition extends Post_Type {
 		// init imports.
 		Imports::get_instance()->init();
 
+		// add our own widgets.
+		Widgets::get_instance()->init();
+
 		// enable extensions.
 		Extensions::get_instance()->init();
 
@@ -145,6 +150,7 @@ class PersonioPosition extends Post_Type {
 		add_filter( 'personio_integration_limit', array( $this, 'set_limit' ), 10, 2 );
 		add_filter( 'personio_integration_light_help_tabs', array( $this, 'add_help' ), 20 );
 		add_filter( 'personio_integration_light_help_tabs', array( $this, 'add_shortcode_help' ), 60 );
+		add_filter( 'personio_integration_light_edit_position_box_taxonomy', array( $this, 'show_hint_for_additional_offices' ), 10, 2 );
 
 		// misc hooks.
 		add_filter( 'posts_search', array( $this, 'extend_search' ), 10, 2 );
@@ -430,178 +436,8 @@ class PersonioPosition extends Post_Type {
 	 * @return string
 	 */
 	public function shortcode_archive( array $attributes = array() ): string {
-		// set pagination settings.
-		$pagination = true;
-		/**
-		 * Set pagination settings.
-		 *
-		 * @since 1.2.0 Available since 1.2.0.
-		 *
-		 * @param bool $pagination The pagination setting (true to disable it).
-		 *
-		 * @noinspection PhpConditionAlreadyCheckedInspection
-		 */
-		$pagination = apply_filters( 'personio_integration_pagination', $pagination );
-
-		// define the default values for each attribute.
-		$attribute_defaults = array(
-			'lang'                    => Languages::get_instance()->get_current_lang(),
-			'showfilter'              => ( 1 === absint( get_option( 'personioIntegrationEnableFilter' ) ) ),
-			'filter'                  => implode( ',', get_option( 'personioIntegrationTemplateFilter' ) ),
-			'filtertype'              => get_option( 'personioIntegrationFilterType' ),
-			'template'                => '',
-			'templates'               => implode( ',', get_option( 'personioIntegrationTemplateContentList' ) ),
-			'listing_template'        => get_option( 'personioIntegrationTemplateContentListingTemplate' ),
-			'excerpt_template'        => get_option( 'personioIntegrationTemplateListingExcerptsTemplate' ),
-			'jobdescription_template' => get_option( 'personioIntegrationTemplateListingContentTemplate' ),
-			'excerpt'                 => implode( ',', get_option( 'personioIntegrationTemplateExcerptDefaults' ) ),
-			'ids'                     => '',
-			'donotlink'               => ( 0 === absint( get_option( 'personioIntegrationEnableLinkInList' ) ) ),
-			'sort'                    => 'asc',
-			'sortby'                  => 'title',
-			'limit'                   => 0,
-			'nopagination'            => $pagination,
-			'groupby'                 => '',
-			'styles'                  => '',
-			'classes'                 => $this->get_default_classes(),
-			'anchor'                  => '',
-			'link_to_anchor'          => '',
-		);
-
-		// define the settings for each attribute (array or string).
-		$attribute_settings = array(
-			'id'                      => 'string',
-			'lang'                    => 'string',
-			'showfilter'              => 'bool',
-			'filter'                  => 'array',
-			'template'                => 'string',
-			'listing_template'        => 'listing_template',
-			'excerpt_template'        => 'excerpt_template',
-			'jobdescription_template' => 'jobdescription_template',
-			'templates'               => 'array',
-			'excerpt'                 => 'array',
-			'ids'                     => 'array',
-			'donotlink'               => 'bool',
-			'sort'                    => 'string',
-			'sortby'                  => 'string',
-			'limit'                   => 'unsignedint',
-			'filtertype'              => 'string',
-			'nopagination'            => 'bool',
-			'groupby'                 => 'string',
-			'styles'                  => 'string',
-			'classes'                 => 'string',
-			'anchor'                  => 'string',
-			'link_to_anchor'          => 'string',
-		);
-
-		// add taxonomies which are available as filter.
-		foreach ( Taxonomies::get_instance()->get_taxonomies() as $taxonomy_name => $taxonomy ) {
-			// bail if no slug is set for this taxonomy.
-			if ( empty( $taxonomy['slug'] ) ) {
-				continue;
-			}
-			if ( ! empty( $GLOBALS['wp']->query_vars['personiofilter'] ) && ! empty( $GLOBALS['wp']->query_vars['personiofilter'][ $taxonomy['slug'] ] ) && ( 1 === absint( $taxonomy['useInFilter'] ) ) ) {
-				$attribute_defaults[ $taxonomy['slug'] ] = 0;
-				$attribute_settings[ $taxonomy['slug'] ] = 'filter';
-			}
-		}
-
-		// get the attributes to filter.
-		$personio_attributes = Helper::get_shortcode_attributes( $attribute_defaults, $attribute_settings, $attributes );
-
-		// get positions-object for search.
-		$positions_obj = Positions::get_instance();
-
-		// filter for specific ids.
-		if ( ! empty( $personio_attributes['ids'][0] ) ) {
-			// convert id-list from PersonioId in post_id.
-			$resulting_list = array();
-			foreach ( $personio_attributes['ids'] as $personio_id ) {
-				$position = $positions_obj->get_position_by_personio_id( $personio_id );
-				if ( $position instanceof Position ) {
-					$resulting_list[] = $position->get_id();
-				}
-			}
-			$personio_attributes['ids'] = $resulting_list;
-		}
-
-		// set limits.
-		$limit_by_wp   = absint( get_option( 'posts_per_page' ) );
-		$limit_by_list = absint( $personio_attributes['limit'] );
-
-		/**
-		 * Change the limit for positions in frontend.
-		 *
-		 * @since 2.0.0 Available since 2.0.0.
-		 *
-		 * @param int $limit_by_wp The limit define by wp which will be used for the list.
-		 * @param int $limit_by_list The limit explicit set for this listing.
-		 */
-		$personio_attributes['limit'] = apply_filters( 'personio_integration_limit', $limit_by_wp, $limit_by_list );
-
-		// get the positions.
-		$positions                         = $positions_obj->get_positions( $personio_attributes['limit'], $personio_attributes );
-		$GLOBALS['personio_query_results'] = $positions_obj->get_results();
-
-		/**
-		 * Change settings for output.
-		 *
-		 * @since 1.2.0 Available since 1.2.0.
-		 *
-		 * @param array $personio_attributes The attributes used for this output.
-		 * @param array $attribute_defaults The default attributes.
-		 */
-		$personio_attributes = apply_filters( 'personio_integration_get_template', $personio_attributes, $attribute_defaults );
-
-		/**
-		 * Run custom actions before the output of the archive listing.
-		 *
-		 * @since 3.2.0 Available since 3.2.0.
-		 * @param array $personio_attributes List of attributes.
-		 */
-		do_action( 'personio_integration_get_template_before', $personio_attributes );
-
-		// generate styling.
-		Helper::add_inline_style( $personio_attributes['styles'] );
-
-		// for backwards compatibility.
-		$form_id = '';
-
-		// set the group-title.
-		$group_title = '';
-
-		// get pagination.
-		$query      = array(
-			'base'    => str_replace( (string) PHP_INT_MAX, '%#%', esc_url( get_pagenum_link( PHP_INT_MAX ) ) ),
-			'format'  => '?paged=%#%',
-			'current' => max( 1, get_query_var( 'paged' ) ),
-			'total'   => $positions_obj->get_results()->max_num_pages,
-		);
-		$pagination = paginate_links( $query );
-		if ( is_null( $pagination ) ) {
-			$pagination = '';
-		}
-
-		// collect the output.
-		ob_start();
-		if ( ! empty( $personio_attributes['styles'] ) && Helper::is_rest_request() ) {
-			wp_styles()->print_inline_style( 'wp-block-library' );
-		}
-
-		// embed filter.
-		require Templates::get_instance()->get_template( 'parts/part-filter.php' );
-
-		// embed the listing content.
-		include Templates::get_instance()->get_template( 'parts/listing.php' );
-
-		// get the content.
-		$content = ob_get_clean();
-
-		// return the content.
-		if ( ! $content ) {
-			return '';
-		}
-		return $content;
+		_deprecated_function( 'PersonioPosition::shortcode_archive()', '5.0.0', __( 'our widget objects', 'personio-integration-light' ) );
+		return Archive::get_instance()->render( $attributes );
 	}
 
 	/**
@@ -1475,8 +1311,9 @@ class PersonioPosition extends Post_Type {
 		 *
 		 * @since 4.1.0 Available since 4.1.0.
 		 * @param Position $position_obj The position as object.
+		 * @param array $attr Attributes used for this meta box.
 		 */
-		do_action( 'personio_integration_light_edit_position_box_taxonomy', $position_obj );
+		do_action( 'personio_integration_light_edit_position_box_taxonomy', $position_obj, $attr );
 	}
 
 	/**
@@ -2101,14 +1938,13 @@ class PersonioPosition extends Post_Type {
 	 * @return array<int,array<string,string>>
 	 */
 	private function get_pro_extensions(): array {
-		$false = false;
+		$false    = false;
 		/**
-		 * Hide the extensions for pro-version.
+		 * Hide hint for Pro-plugin.
 		 *
 		 * @since 3.0.0 Available since 3.0.0
 		 *
-		 * @param bool $false Set true to hide the extensions.
-		 *
+		 * @param bool $false Set true to hide the hint.
 		 * @noinspection PhpConditionAlreadyCheckedInspection
 		 */
 		if ( apply_filters( 'personio_integration_hide_pro_hints', $false ) ) {
@@ -2297,13 +2133,14 @@ class PersonioPosition extends Post_Type {
 		/* translators: %1$s will be replaced by a URL. */
 		$content .= '<li>' . sprintf( __( 'Change settings <a href="%1$s">for the template</a> to optimize the view.', 'personio-integration-light' ), esc_url( Helper::get_settings_url( 'personioPositions', 'templates' ) ) ) . '</li>';
 		// add menu entry for applications (with hint to pro).
-		$false = false;
+		$false    = false;
 		/**
-		 * Hide pro hint in help.
+		 * Hide hint for Pro-plugin.
 		 *
 		 * @since 3.0.0 Available since 3.0.0
 		 *
-		 * @param bool $false Set true to hide the buttons.
+		 * @param bool $false Set true to hide the hint.
+		 * @noinspection PhpConditionAlreadyCheckedInspection
 		 */
 		if ( ! apply_filters( 'personio_integration_hide_pro_hints', $false ) ) {
 			/* translators: %1$s will be replaced by a URL. */
@@ -2336,11 +2173,12 @@ class PersonioPosition extends Post_Type {
 		$content .= '<p>' . __( 'Shortcodes should only be used if your own theme or PageBuilder do not offer any other options.', 'personio-integration-light' ) . '</p>';
 		$false    = false;
 		/**
-		 * Hide pro hint in help.
+		 * Hide hint for Pro-plugin.
 		 *
 		 * @since 3.0.0 Available since 3.0.0
 		 *
-		 * @param bool $false Set true to hide the buttons.
+		 * @param bool $false Set true to hide the hint.
+		 * @noinspection PhpConditionAlreadyCheckedInspection
 		 */
 		if ( ! apply_filters( 'personio_integration_hide_pro_hints', $false ) ) {
 			$content .= '<p>' . __( 'With <i>Personio Integration Light</i>, we only support the Block Editor in this regard.', 'personio-integration-light' ) . '</p>';
@@ -2365,11 +2203,12 @@ class PersonioPosition extends Post_Type {
 		$content .= '<li>' . __( 'If you want to customize the output of shortcodes in terms of styling, you have to write and store the necessary style properties yourself. If necessary, contact the person responsible for your project.', 'personio-integration-light' ) . '</li>';
 		$false    = false;
 		/**
-		 * Hide pro hint in help.
+		 * Hide hint for Pro-plugin.
 		 *
 		 * @since 3.0.0 Available since 3.0.0
 		 *
-		 * @param bool $false Set true to hide the buttons.
+		 * @param bool $false Set true to hide the hint.
+		 * @noinspection PhpConditionAlreadyCheckedInspection
 		 */
 		if ( ! apply_filters( 'personio_integration_hide_pro_hints', $false ) ) {
 			/* translators: %1$s will be replaced by a URL. */
@@ -2386,37 +2225,6 @@ class PersonioPosition extends Post_Type {
 
 		// return resulting list.
 		return $help_list;
-	}
-
-	/**
-	 * Return list of default classes depending on main settings.
-	 *
-	 * @return string
-	 */
-	private function get_default_classes(): string {
-		// initiate the list of classes.
-		$css_classes = array();
-
-		// add hide title.
-		if ( 1 === absint( get_option( 'personioIntegrationHideFilterTitle' ) ) ) {
-			$css_classes[] = 'personio-hide-title';
-		}
-
-		// add hide reset.
-		if ( 1 === absint( get_option( 'personioIntegrationHideFilterReset' ) ) ) {
-			$css_classes[] = 'personio-hide-reset';
-		}
-
-		/**
-		 * Filter the default classes for each output of positions.
-		 *
-		 * @since 4.2.0 Available since 4.2.0
-		 * @param array<int,string> $css_classes List of classes.
-		 */
-		$css_classes = apply_filters( 'personio_integration_light_default_css_classes', $css_classes );
-
-		// return the resulting list.
-		return implode( ' ', $css_classes );
 	}
 
 	/**
@@ -2496,5 +2304,39 @@ class PersonioPosition extends Post_Type {
 
 		// return an empty list.
 		return array();
+	}
+
+	/**
+	 * Show hint about additional offices, usable in Pro-plugin.
+	 *
+	 * @param Position $position_obj The called position object.
+	 * @param array    $attr The attributes of the meta box.
+	 *
+	 * @return void
+	 * @noinspection PhpUnusedParameterInspection
+	 */
+	public function show_hint_for_additional_offices( Position $position_obj, array $attr ): void {
+		$false    = false;
+		/**
+		 * Hide hint for Pro-plugin.
+		 *
+		 * @since 3.0.0 Available since 3.0.0
+		 *
+		 * @param bool $false Set true to hide the hint.
+		 * @noinspection PhpConditionAlreadyCheckedInspection
+		 */
+		if ( apply_filters( 'personio_integration_hide_pro_hints', $false ) ) {
+			return;
+		}
+
+		// get the requested taxonomy from the box ID as string.
+		$taxonomy_name = str_replace( $this->get_name() . '-taxonomy-', '', $attr['id'] );
+
+		// bail if taxonomy name is not office.
+		if( $taxonomy_name !== WP_PERSONIO_INTEGRATION_TAXONOMY_OFFICE ) {
+			return;
+		}
+
+		echo wp_kses_post( Admin::get_instance()->get_pro_hint( __( 'Use additional offices with %1$s', 'personio-integration-light' ) ) );
 	}
 }