+++ blocks/details/src/edit.js
@@ -62,7 +62,7 @@ export default function Edit( object ) {
       ]);
     }, []);
     templates = useSelect((select) => {
-        return select('core').getEntityRecords('personio/v1', 'details-templates') || [];
+        return select('core').getEntityRecords('personio/v1', 'details-templates', { per_page: 20 }) || [];
       }
     );
   }
@@ -80,7 +80,7 @@ export default function Edit( object ) {
       ]);
     }, []);
     personioTaxonomies = useSelect((select) => {
-        return select('core').getEntityRecords('personio/v1', 'taxonomies') || [];
+        return select('core').getEntityRecords('personio/v1', 'taxonomies', { per_page: 20 }) || [];
       }
     );
   }